requirements.txt
Flask
flask-socketio
flask_session 
flask_sqlalchemy
python-dotenv
SQLAlchemy
psycopg2-binary
gunicorn
requests